requirements.txt
aioconsole==0.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chess==1.5.0
colorama==0.4.4
greenlet==0.4.13
idna==2.10
lru-dict==1.1.7
pytz==2021.1
readline==6.2.4.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
vmprof==0.4.15
